Testing shifting in instruction set, this may be unessecary. If I got 5 bits, -1 would represent 11111 or 0x1F. I thought 0 could be -1 and 1 be 0. It wouldn't really matter here, but I need to rethink this a little bit. How can I make the assembler as general as possible.

+++ parse.h
@@ -85,7 +85,7 @@ typedef struct
 
   unsigned int arg_overflow_len;
   char arg_overflow[MAX_OP_DESC];
-  long int shift;
+  int shift;
   regex_t reg;
 }argument;
